Start reorganizing flightsql
